refactor: windowSize 커스텀 훅 분리
