+++ src/hooks/useWindowSize.ts
@@ -1,0 +1,20 @@
+import { useEffect, useState } from 'react';
+
+const useWindowSize = () => {
+  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowWidth(window.innerWidth);
+    };
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return { windowWidth };
+};
+
+export default useWindowSize;
